+++ Addons/Schumix.CalendarAddon/Commands/BanCommand.cs
@@ -19,6 +19,7 @@
  */
 
 using System;
+using System.Text.RegularExpressions;
 using Schumix.Api.Irc;
 using Schumix.Irc;
 using Schumix.Irc.Commands;
@@ -34,6 +35,7 @@ namespace Schumix.CalendarAddon.Commands
 		private readonly LocalizationManager sLManager = Singleton<LocalizationManager>.Instance;
 		private readonly Utilities sUtilities = Singleton<Utilities>.Instance;
 		private readonly IrcBase sIrcBase = Singleton<IrcBase>.Instance;
+		private Regex _regex;
 		private Unban sUnban;
 		private Ban sBan;
 
@@ -41,6 +43,47 @@ namespace Schumix.CalendarAddon.Commands
 		{
 			sBan = new Ban(ServerName);
 			sUnban = new Unban(ServerName);
+			_regex = new Regex(@"((?<year>[0-9]{1,4})(?:[\.\s]+))?"                         // Year
+			                   + @"((?<month>[0-9]{1,2}|[a-zóüöúőűáéí]{3,20})(?:[\.\s]+))?" // Month
+			                   + @"((?<day>[0-9]{1,2})(?:[\.\s]+))?"                        // Day
+			                   + @"(?<hour>[0-9]{1,2})(?:[:])?"                             // Hour
+			                   + @"(?<minute>[0-9]{1,2})?"                                  // Minute
+			                   + @"((?:[\s]+)(?<text>(.*)))?", RegexOptions.Compiled | RegexOptions.IgnoreCase);
+		}
+
+		private int GetYear(string args)
+		{
+			return _regex.IsMatch(args) ? _regex.Match(args).Groups["year"].ToString().ToNumber(-1).ToInt() : -1;
+		}
+
+		private string GetMonth(string args)
+		{
+			return _regex.IsMatch(args) ? _regex.Match(args).Groups["month"].ToString() : string.Empty;
+		}
+
+		private int GetDay(string args)
+		{
+			return _regex.IsMatch(args) ? _regex.Match(args).Groups["day"].ToString().ToNumber(32).ToInt() : 32;
+		}
+
+		private int GetHour(string args)
+		{
+			return _regex.IsMatch(args) ? _regex.Match(args).Groups["hour"].ToString().ToNumber(25).ToInt() : 25;
+		}
+
+		private int GetMinute(string args)
+		{
+			return _regex.IsMatch(args) ? _regex.Match(args).Groups["minute"].ToString().ToNumber(61).ToInt() : 61;
+		}
+
+		private string GetMessage(string args)
+		{
+			return _regex.IsMatch(args) ? _regex.Match(args).Groups["text"].ToString() : string.Empty;
+		}
+
+		private bool IsHourAndMinute(string args)
+		{
+			return GetYear(args) == -1 && GetMonth(args) == string.Empty && GetDay(args) == 32;
 		}
 
 		public void HandleBan(IRCMessage sIRCMessage)
@@ -62,29 +105,31 @@ namespace Schumix.CalendarAddon.Commands
 				return;
 			}
 
-			if(sIRCMessage.Info[5].Contains(SchumixBase.Colon.ToString()))
+			string args = sIRCMessage.Info.SplitToString(5, SchumixBase.Space);
+
+			if(IsHourAndMinute(args))
 			{
-				if(sIRCMessage.Info.Length < 7)
+				if(GetMessage(args).IsNullOrEmpty())
 				{
 					sSendMessage.SendChatMessage(sIRCMessage, sLManager.GetWarningText("NoReason", sIRCMessage.Channel, sIRCMessage.ServerName));
 					return;
 				}
 
-				int hour = sIRCMessage.Info[5].Substring(0, sIRCMessage.Info[5].IndexOf(SchumixBase.Colon)).ToNumber(25).ToInt();
+				int hour = GetHour(args);
 				if(hour >= 24 || hour < 0)
 				{
 					sSendMessage.SendChatMessage(sIRCMessage, sLManager.GetWarningText("ErrorHour", sIRCMessage.Channel, sIRCMessage.ServerName));
 					return;
 				}
 
-				int minute = sIRCMessage.Info[5].Substring(sIRCMessage.Info[5].IndexOf(SchumixBase.Colon)+1).ToNumber(61).ToInt();
+				int minute = GetMinute(args);
 				if(minute >= 60 || minute < 0)
 				{
 					sSendMessage.SendChatMessage(sIRCMessage, sLManager.GetWarningText("ErrorMinute", sIRCMessage.Channel, sIRCMessage.ServerName));
 					return;
 				}
 
-				sSendMessage.SendChatMessage(sIRCMessage, sBan.BanName(sIRCMessage.Info[4].ToLower(), sIRCMessage.Channel, sIRCMessage.Info.SplitToString(6, SchumixBase.Space), hour, minute));
+				sSendMessage.SendChatMessage(sIRCMessage, sBan.BanName(sIRCMessage.Info[4].ToLower(), sIRCMessage.Channel, GetMessage(args), hour, minute));
 			}
 			else
 			{
@@ -94,55 +139,48 @@ namespace Schumix.CalendarAddon.Commands
 					return;
 				}
 
-				if(sIRCMessage.Info.Length < 8)
+				if(GetMessage(args).IsNullOrEmpty())
 				{
 					sSendMessage.SendChatMessage(sIRCMessage, sLManager.GetWarningText("NoReason", sIRCMessage.Channel, sIRCMessage.ServerName));
 					return;
 				}
 
-				string[] s = sIRCMessage.Info[5].Split(SchumixBase.Point);
-				if(s.Length < 3)
-				{
-					sSendMessage.SendChatMessage(sIRCMessage, sLManager.GetCommandText("ban", sIRCMessage.Channel, sIRCMessage.ServerName));
-					return;
-				}
-
-				int year = s[0].ToNumber(-1).ToInt();
+				int year = GetYear(args);
 				if(year < 0)
 				{
 					sSendMessage.SendChatMessage(sIRCMessage, sLManager.GetWarningText("ErrorYear", sIRCMessage.Channel, sIRCMessage.ServerName));
 					return;
 				}
 
-				int month = s[1].ToNumber(13).ToInt();
+				int month = GetMonth(args).GetMonthNameInInt(sLManager.GetChannelLocalization(sIRCMessage.Channel, sIRCMessage.ServerName));
 				if(month > 12 || month <= 0)
 				{
 					sSendMessage.SendChatMessage(sIRCMessage, sLManager.GetWarningText("ErrorMonth", sIRCMessage.Channel, sIRCMessage.ServerName));
 					return;
 				}
 
-				int day = s[2].ToNumber(32).ToInt();
+				int day = GetDay(args);
 				if(!sUtilities.IsDay(year, month, day))
 				{
 					sSendMessage.SendChatMessage(sIRCMessage, sLManager.GetWarningText("ErrorDay", sIRCMessage.Channel, sIRCMessage.ServerName));
 					return;
 				}
 
-				int hour = sIRCMessage.Info[6].Substring(0, sIRCMessage.Info[6].IndexOf(SchumixBase.Colon)).ToNumber(25).ToInt();
+				int hour = GetHour(args);
 				if(hour >= 24 || hour < 0)
 				{
 					sSendMessage.SendChatMessage(sIRCMessage, sLManager.GetWarningText("ErrorHour", sIRCMessage.Channel, sIRCMessage.ServerName));
 					return;
 				}
 
-				int minute = sIRCMessage.Info[6].Substring(sIRCMessage.Info[6].IndexOf(SchumixBase.Colon)+1).ToNumber(61).ToInt();
+				int minute = GetMinute(args);
 				if(minute >= 60 || minute < 0)
 				{
 					sSendMessage.SendChatMessage(sIRCMessage, sLManager.GetWarningText("ErrorMinute", sIRCMessage.Channel, sIRCMessage.ServerName));
 					return;
 				}
 
-				sSendMessage.SendChatMessage(sIRCMessage, sBan.BanName(sIRCMessage.Info[4].ToLower(), sIRCMessage.Channel, sIRCMessage.Info.SplitToString(7, SchumixBase.Space), year, month, day, hour, minute));
+				sSendMessage.SendChatMessage(sIRCMessage, sBan.BanName(sIRCMessage.Info[4].ToLower(), sIRCMessage.Channel, GetMessage(args), year, month, day, hour, minute));
 			}
 		}