+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## 4.2.0 (2013-xx-xx)
 
+ * Calendar addonban a calendar és a ban parancsoknál alkalmazva az új regex-es működés.
  * IsValueBiggerDateTimeNow() függvény kijavítva.
  * Kick parancs feljavítása elkészült.
  * LogToFile függvény átnevezve LogInFile függvényre.